mbbsd: refactor ::AccountRecovery::EmailCodeChallenge to EmailCodeChallenge
Currently we verify email (contact-email or register-email as valid email) again
only in AccountRecovery.

Refactor ::AccountRecovery::EmailCodeChallenge to EmailCodeChallenge
and the corresponding helper static functions to prepare for the extension
to verify email in other sensitive scenarios.
