+++ mbbsd/recover.cc
@@ -11,19 +11,14 @@ extern "C" {
 
 #include <optional>
 #include <string>
+#include <vector>
 
-#include "verifydb.h"
-#include "verifydb.fbs.h"
+#include "user_handle.hpp"
+#include "email_challenge.hpp"
 
 namespace {
 
 constexpr int kMaxErrCnt = 3;
-constexpr size_t kCodeLen = 30;
-
-struct UserHandle {
-  std::string userid;
-  int64_t generation;
-};
 
 class AccountRecovery {
 public:
@@ -36,18 +31,12 @@ private:
   int y_ = 2;
 
   bool LoadUser(const char *userid);
-  bool LoadVerifyDbEmail();
   void InputUserEmail();
-  void EmailCodeChallenge();
   void ResetPasswd();
 
-  static std::optional<std::string> NormalizeEmail(const std::string &email);
-  static bool SendRecoveryCode(const std::string &email,
-                               const std::string &code);
   static bool SetPasswd(const UserHandle &user, const char *hashed_passwd);
   static void LogToSecurity(const UserHandle &user, const std::string &email);
   static void NotifyUser(const std::string &userid, const std::string &email);
-  static std::string GenCode(size_t len);
   static void UserErrorExit();
 };
 
@@ -68,56 +57,8 @@ bool AccountRecovery::LoadUser(const char *userid) {
   user.generation = u.firstlogin;
   user_ = user;
 
-#ifdef USEREC_EMAIL_IS_CONTACT
-  auto email = NormalizeEmail(u.email);
-  if (email)
-    all_emails_.push_back(email.value());
-#endif
+  LoadUserEmail(u, all_emails_);
 
-  return true;
-}
-
-bool AccountRecovery::LoadVerifyDbEmail() {
-  Bytes buf;
-  const VerifyDb::GetReply *reply;
-  if (!verifydb_getuser(user_->userid.c_str(), user_->generation, &buf, &reply))
-    return false;
-
-  if (reply->entry()) {
-    for (const auto *ent : *reply->entry()) {
-      if (ent->vmethod() == VMETHOD_EMAIL && ent->vkey() != nullptr) {
-        auto email = NormalizeEmail(ent->vkey()->str());
-        if (email)
-          all_emails_.push_back(email.value());
-      }
-    }
-  }
-  return true;
-}
-
-// static
-std::optional<std::string>
-AccountRecovery::NormalizeEmail(const std::string &email) {
-  auto out = email;
-  for (auto& c : out)
-    c = std::tolower(c);
-  auto idx = out.find('@');
-  if (idx == std::string::npos)
-    return std::nullopt;
-  if (idx != out.rfind('@'))
-    return std::nullopt;
-  return out;
-}
-
-// static
-bool AccountRecovery::SendRecoveryCode(const std::string &email,
-                                       const std::string &code) {
-  std::string subject;
-  subject.append(" " BBSNAME " - 重設密碼認證碼 [ ");
-  subject.append(code);
-  subject.append(" ] @ IP ");
-  subject.append(fromhost);
-  bsmtp("etc/recovermail", subject.c_str(), email.c_str(), "non-exist");
   return true;
 }
 
@@ -155,13 +96,6 @@ void AccountRecovery::LogToSecurity(const UserHandle &user,
   msg.append("\n");
 
   post_msg(BN_SECURITY, title.c_str(), msg.c_str(), "[系統安全局]");
-}
-
-// static
-std::string AccountRecovery::GenCode(size_t len) {
-  std::string s(len, '\0');
-  random_text_code(&s[0], s.size());
-  return s;
 }
 
 // static
@@ -209,58 +143,9 @@ void AccountRecovery::InputUserEmail() {
   move(y_, 0);
   clrtoeol(); // There might be error message at this line.
 
-  if (!LoadVerifyDbEmail()) {
+  if (!LoadVerifyDbEmail(user_, all_emails_)) {
     vmsg("系統錯誤，請稍候再試。");
     exit(0);
-  }
-}
-
-void AccountRecovery::EmailCodeChallenge() {
-  // Generate code.
-  auto code = GenCode(kCodeLen);
-
-  y_++;
-  mvprints(y_, 0, "系統處理中...");
-  doupdate();
-
-  // Check and send recovery code. Don't exit if email doesn't match, so user
-  // can't guess email.
-  bool email_matches = false;
-  for (const auto &email : all_emails_) {
-    if (email == email_)
-      email_matches = true;
-  }
-
-  if (email_matches) {
-    // We only want to send email if the user input the matching address.
-    // Silently fail if email is not matching, so that user cannot guess other
-    // user's email address.
-    SendRecoveryCode(email_, code);
-  }
-  // Add a random 5-10s delay to prevent timing oracle.
-  usleep(5000000 + random() % 5000000);
-  mvprints(y_++, 0, "若您輸入的資料正確，系統已將認證碼寄送至您的信箱。");
-
-  // Input code.
-  char incode[kCodeLen + 1] = {};
-  int errcnt = 0;
-  while (1) {
-    getdata_buf(y_, 0, "請收信後輸入認證碼：", incode, sizeof(incode), DOECHO);
-    if (email_matches && code == std::string(incode))
-      break;
-    if (++errcnt >= kMaxErrCnt)
-      UserErrorExit();
-    mvprints(y_ + 1, 0, "認證碼錯誤！認證碼共有 %d 字元。", (int)kCodeLen);
-    incode[0] = '\0';
-  }
-  y_++;
-  move(y_, 0);
-  clrtoeol(); // There might be error message at this line.
-
-  // Some paranoid checkings, we are about to let user reset their password.
-  if (code.size() != kCodeLen || code != std::string(incode)) {
-    assert(false);
-    exit(1);
   }
 }
 
@@ -325,7 +210,7 @@ void AccountRecovery::ResetPasswd() {
 void AccountRecovery::Run() {
   vs_hdr("取回帳號");
   InputUserEmail();
-  EmailCodeChallenge();
+  EmailCodeChallenge(email_, all_emails_, y_);
   ResetPasswd();
 }
